fix(rag): correct score decode types from measured pg_typeof + bump v0.1.7
Measured against live Postgres:
- '1 - (vector <=> vector)' returns double precision (FLOAT8) -> decode f64
- 'ts_rank(...)' returns real (FLOAT4) -> decode f32

v0.1.5/v0.1.6 guessed the types and caused repeated ColumnDecode panics.
This fix is grounded in pg_typeof evidence, not assumption.

+++ pacgate-ai/crates/pacgate-rag/src/lib.rs
@@ -143,7 +143,7 @@ impl RagStore {
         // Build query with optional jurisdiction + source_level filters
         let (sql, jur_param, sl_param) = Self::build_filtered_sql(
             "SELECT c.content, c.page, d.name as doc_name,
-                    1 - (c.embedding <=> $3::vector) as score
+                    (1 - (c.embedding <=> $3::vector))::float8 as score
              FROM kb_chunks c
              JOIN documents d ON c.document_id = d.id
              WHERE c.tenant_id = $1 AND c.matter_id = $2
@@ -172,9 +172,8 @@ impl RagStore {
             .into_iter()
             .map(|row| KbSearchResult {
                 content: row.get("content"),
-                // pgvector's `<=>` returns FLOAT4 (real), not FLOAT8. Decode as
-                // f32 to match the SQL type and avoid a ColumnDecode panic.
-                score: row.get::<f32, _>("score"),
+                // Score is cast to ::float8 in the SQL so it is always FLOAT8.
+                score: row.get::<f64, _>("score") as f32,
                 source_doc: row.get("doc_name"),
                 page: row.get::<Option<i32>, _>("page").map(|p| p as u32),
             })
@@ -221,9 +220,8 @@ impl RagStore {
             .into_iter()
             .map(|row| KbSearchResult {
                 content: row.get("content"),
-                // ts_rank returns FLOAT8 (double precision), not FLOAT4. Decode as
-                // f64 to match the SQL type and avoid a ColumnDecode panic.
-                score: row.get::<f64, _>("score") as f32,
+                // ts_rank returns real (FLOAT4), not FLOAT8. Decode as f32.
+                score: row.get::<f32, _>("score"),
                 source_doc: row.get("doc_name"),
                 page: row.get::<Option<i32>, _>("page").map(|p| p as u32),
             })
